First two segments with powerfont separator.
Add new configurations for the height, position, the font and segment
separators.
For the font, a fork of the original lemonbar has been used, to use Xtf
fonts.
The font size is set in relation to the bar height by default.
The bar display two example segments on the right with the separators.