Individual segment definition possible.
Introduce the initSegments function which is aware of the construct of
all segments and spawn the processes, which update the segments
periodically.
Introduce the updateSegment function, which is responsible for a single
segment, update its content and build the format string, passed to the
fifo.
The reading function is now updating individual segment and pass the
concatenation of the whole format string composition to the bar.