Split segment implementation and configuration.
Export customizable segment variables into a separated configuration
file for each segment.
This allows a more flexible usage and overwriting by the users custom
configuration file.
Such a configuration file must not exist!
Extend the SegmentFinder to differ between the implementation and the
configuration of a segment.
Adjust the whole chain to the SegmentHandler to use the configurations.
Adjust all default segments to use such configuration file.
Extend the inline documentation of the segments.