Fix update cycle interval by segment definition.
The bug was, that the update cycle for the SegmentHandler has been
determined by the main script, but the segments configuration has been
loaded after that in the SegmentHandler itself.
By this the segments variable definition wasn't used.
The SegmentHandler does this determination on his own now.
Add missing '%' signs for battery and sound segment.