Introduce subscription segments.
The SegmentHandler differs between cycle based and subscription
segments.
Define a core function that concatenate the static separators with the
current state and forward it to the main FIFO.
Define a function which will handle the whole process depending on the
segment type.
Extend the default configuration for the prefix, a segment FIFO uses.
Implement the function to handle the segment subscription.
Extend the interface of segments for the subscription to have an init
function and the formatting.