Introduce the abbreviation utilities.
Provide the functionality for all segments to abbreviate their content
on the users wishes.
An segment has to specifically mark abbreviations as available by its
configuration file.
The SegmentHandler is responsible to load the utilities for its segment.
If abbreviation is available it can be enabled or disabled.
On the high level this can be en-/disabled per segment by the users
custom configuration.
If such specific option is not defined, the overall enable variable is
checked.
Last named variable is defined in the default configuration and can be
overwritten by the users custom configuration.
Abbreviations comes with three more variables to configure.
First there is the length off that length a content string is allowed to
be in maximum.
Next is the abbreviation string which will be used as replacement for
the cut out string part.
At least the type defines where the string gets cut off.
All these three variables are also defined on the default layer and are
customizable by the user, also per segment.